+++ src/plugin/ffi.rs
@@ -479,7 +479,7 @@ pub extern "C" fn pico_ffi_background_set_jobs_shutdown_timeout(
     0
 }
 
-/// See [`crate::auth::authenticate`] for more information.
+/// See [`crate::auth::authenticate_with_password`] for more information.
 #[no_mangle]
 #[sabi_extern_fn]
 pub extern "C" fn pico_ffi_authenticate(name: FfiSafeStr, password: FfiSafeBytes) -> i32 {
@@ -487,7 +487,7 @@ pub extern "C" fn pico_ffi_authenticate(name: FfiSafeStr, password: FfiSafeBytes
     let name = unsafe { name.as_str() };
     let password = unsafe { password.as_bytes() };
 
-    match crate::auth::authenticate(name, password, None) {
+    match crate::auth::authenticate_with_password(name, password) {
         Ok(_) => 0,
         Err(e) => {
             e.into_box_error().set_last();